+++ lib/masterdb.ml
@@ -38,12 +38,7 @@ type elt_t = {
   title : string;
   objective : string;
   project : string;
-  schedule : string option;
-  lead : string;
   team : string;
-  category : string;
-  links : string option;
-  reports : string list;
   status : status_t option;
 }
 
@@ -87,15 +82,6 @@ let load_csv ?(separator = ',') f =
       ~f:(fun row ->
         line := !line + 1;
         let find_and_trim col = Csv.Row.find row col |> String.trim in
-        let find_and_trim_opt col =
-          let v = find_and_trim col in
-          if v <> "" then Some v else None
-        in
-        let find_and_trim_list col =
-          match find_and_trim_opt col with
-          | None -> []
-          | Some s -> Str.split (Str.regexp "[ ,]+") s
-        in
         let printable_id = find_and_trim "id" in
         let e =
           {
@@ -104,12 +90,7 @@ let load_csv ?(separator = ',') f =
             title = find_and_trim "title" |> normalise_title;
             objective = find_and_trim "objective";
             project = find_and_trim "project";
-            schedule = find_and_trim_opt "schedule";
-            lead = find_and_trim "lead";
             team = find_and_trim "team";
-            category = find_and_trim "category";
-            reports = find_and_trim_list "reports";
-            links = find_and_trim_opt "links";
             status = find_and_trim "status" |> status_of_string;
           }
         in
@@ -151,22 +132,4 @@ let filter_krs t f =
 let find_krs_for_teams t teams =
   let teams = List.map String.uppercase_ascii teams in
   let p e = List.exists (String.equal (String.uppercase_ascii e.team)) teams in
-  filter_krs t p
-
-let find_krs_for_categories t categories =
-  let categories = List.map String.uppercase_ascii categories in
-  let p e =
-    List.exists (String.equal (String.uppercase_ascii e.category)) categories
-  in
-  filter_krs t p
-
-let find_krs_for_reports t reports =
-  let reports = List.map String.uppercase_ascii reports in
-  let p e =
-    List.exists
-      (fun report ->
-        let report = String.uppercase_ascii report in
-        List.exists (String.equal report) reports)
-      e.reports
-  in
   filter_krs t p